turned ttkUtils into class
