+++ core/vtk/ttkAlgorithm/ttk.module
@@ -2,6 +2,7 @@ NAME
   ttkAlgorithm
 SOURCES
   ttkAlgorithm.cpp
+  ttkUtils.cpp
 HEADERS
   ttkAlgorithm.h
   ttkUtils.h
